+++ src/com/kaylerrenslow/armaDialogCreator/gui/img/ADCImages.java
@@ -14,4 +14,6 @@ public class ADCImages {
 	public static final Image ICON_REFRESH = new Image(ADCImagePaths.ICON_REFRESH);
 	public static final Image ICON_SAVE = new Image(ADCImagePaths.ICON_SAVE);
 	public static final Image ICON_EXPORT = new Image(ADCImagePaths.ICON_EXPORT);
+	public static final Image ICON_PLUS = new Image(ADCImagePaths.ICON_PLUS);
+	public static final Image ICON_MINUS = new Image(ADCImagePaths.ICON_MINUS);
 }