+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/actions/mainMenu/edit/EditRedoAction.java
@@ -3,6 +3,7 @@ package com.kaylerrenslow.armaDialogCreator.gui.main.actions.mainMenu.edit;
 import com.kaylerrenslow.armaDialogCreator.data.ChangeUpdateFailedException;
 import com.kaylerrenslow.armaDialogCreator.data.Changelog;
 import com.kaylerrenslow.armaDialogCreator.data.ChangelogUpdate;
+import com.kaylerrenslow.armaDialogCreator.gui.notification.BoxNotification;
 import com.kaylerrenslow.armaDialogCreator.gui.notification.Notification;
 import com.kaylerrenslow.armaDialogCreator.gui.notification.Notifications;
 import com.kaylerrenslow.armaDialogCreator.main.ExceptionHandler;
@@ -59,7 +60,7 @@ public class EditRedoAction implements EventHandler<ActionEvent> {
 			changelog.redo();
 		} catch (Exception e) {
 			if (e instanceof ChangeUpdateFailedException) {
-				Notifications.showNotification(new Notification(
+				Notifications.showNotification(new BoxNotification(
 								Lang.ApplicationBundle().getString("Notifications.RedoOperation.notification_title"),
 								String.format(Lang.ApplicationBundle().getString("Notifications.RedoOperation.notification_body_f"), e.getMessage()),
 								Notification.DEFAULT_DURATION,