+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/actions/mainMenu/file/FileSaveAction.java
@@ -2,6 +2,7 @@ package com.kaylerrenslow.armaDialogCreator.gui.main.actions.mainMenu.file;
 
 import com.kaylerrenslow.armaDialogCreator.data.ApplicationDataManager;
 import com.kaylerrenslow.armaDialogCreator.data.Project;
+import com.kaylerrenslow.armaDialogCreator.gui.notification.BoxNotification;
 import com.kaylerrenslow.armaDialogCreator.gui.notification.Notification;
 import com.kaylerrenslow.armaDialogCreator.gui.notification.Notifications;
 import com.kaylerrenslow.armaDialogCreator.main.Lang;
@@ -22,14 +23,14 @@ public class FileSaveAction implements EventHandler<ActionEvent> {
 		try {
 			ApplicationDataManager.getInstance().saveProject();
 			File saveFile = Project.getCurrentProject().getProjectSaveDirectory();
-			resultNotification = new Notification(
+			resultNotification = new BoxNotification(
 					Lang.ApplicationBundle().getString("Notifications.ProjectSave.Success.notification_title"),
 					String.format(Lang.ApplicationBundle().getString("Notifications.ProjectSave.Success.notification_body_f"), saveFile.getAbsolutePath())
 			);
 		} catch (IOException e) {
 			e.printStackTrace();
 			String reason = e.getMessage() != null && e.getMessage().length() > 0 ? e.getMessage() : Lang.ApplicationBundle().getString("Notifications.ProjectSave.Fail.unknown_reason");
-			resultNotification = new Notification(
+			resultNotification = new BoxNotification(
 					Lang.ApplicationBundle().getString("Notifications.ProjectSave.Fail.notification_title"),
 					String.format(Lang.ApplicationBundle().getString("Notifications.ProjectSave.Fail.notification_body_f"), reason),
 					10 * 1000, true