+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/popup/NameInputDialog.java
@@ -2,6 +2,7 @@ package com.kaylerrenslow.armaDialogCreator.gui.main.popup;
 
 import com.kaylerrenslow.armaDialogCreator.gui.popup.StageDialog;
 import com.kaylerrenslow.armaDialogCreator.main.ArmaDialogCreator;
+import javafx.beans.property.StringProperty;
 import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.VBox;
@@ -14,7 +15,7 @@ import org.jetbrains.annotations.Nullable;
  @author Kayler
  @since 12/24/2016 */
 public class NameInputDialog extends StageDialog<VBox> {
-	protected final TextField tf = new TextField();
+	protected final TextField textField = new TextField();
 
 	public NameInputDialog(@NotNull String title, @NotNull String message) {
 		this(title, message, null);
@@ -24,19 +25,29 @@ public class NameInputDialog extends StageDialog<VBox> {
 		super(ArmaDialogCreator.getPrimaryStage(), new VBox(5), title, true, true, false);
 		myRootElement.setFillWidth(true);
 		myRootElement.getChildren().add(new Label(message));
-		myRootElement.getChildren().add(tf);
+		myRootElement.getChildren().add(textField);
 		myRootElement.setPrefWidth(320);
 		myStage.setResizable(false);
-		tf.setPromptText(promptText);
-		tf.setFocusTraversable(false);
+		textField.setPromptText(promptText);
+		textField.setFocusTraversable(false);
 	}
 
 	public void setInputText(@Nullable String text) {
-		tf.setText(text);
+		textField.setText(text);
 	}
 
 	@Nullable
 	public String getInputText() {
-		return tf.getText();
+		return textField.getText();
+	}
+
+	@NotNull
+	public TextField getTextField() {
+		return textField;
+	}
+
+	@NotNull
+	public StringProperty inputTextProperty() {
+		return textField.textProperty();
 	}
 }