+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/stringtable/NewStringTableKeyDialog.java
@@ -1,0 +1,101 @@
+package com.kaylerrenslow.armaDialogCreator.gui.main.stringtable;
+
+import com.kaylerrenslow.armaDialogCreator.arma.stringtable.KnownLanguage;
+import com.kaylerrenslow.armaDialogCreator.arma.stringtable.Language;
+import com.kaylerrenslow.armaDialogCreator.arma.stringtable.StringTable;
+import com.kaylerrenslow.armaDialogCreator.arma.stringtable.StringTableKey;
+import com.kaylerrenslow.armaDialogCreator.arma.stringtable.impl.StringTableKeyImpl;
+import com.kaylerrenslow.armaDialogCreator.gui.fxcontrol.CheckMenuButton;
+import com.kaylerrenslow.armaDialogCreator.gui.main.popup.NameInputDialog;
+import com.kaylerrenslow.armaDialogCreator.main.Lang;
+import javafx.beans.value.ChangeListener;
+import javafx.beans.value.ObservableValue;
+import javafx.collections.FXCollections;
+import javafx.collections.ListChangeListener;
+import javafx.geometry.Pos;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.ResourceBundle;
+
+/**
+ @author Kayler
+ @since 12/26/2016 */
+class NewStringTableKeyDialog extends NameInputDialog {
+
+	private final StringTableKey key = new StringTableKeyImpl("str_", FXCollections.observableHashMap());
+	private final StringTable table;
+
+	public NewStringTableKeyDialog(@NotNull StringTable table) {
+		super(Lang.ApplicationBundle().getString("Popups.StringTable.Tab.Edit.insert_key_tooltip"), Lang.ApplicationBundle().getString("Popups.StringTable.Tab.Edit.insert_key_popup_body"));
+		this.table = table;
+		ResourceBundle bundle = Lang.ApplicationBundle();
+
+		setInputText(key.getId());
+
+		CheckMenuButton<Language> menuButtonLangs = new CheckMenuButton<>(bundle.getString("Popups.StringTable.Tab.Edit.insert_key_popup_lang_button_text"), null, KnownLanguage.values());
+		menuButtonLangs.setAlignment(Pos.CENTER_LEFT);
+		menuButtonLangs.getSelectedItems().addListener(new ListChangeListener<Language>() {
+			@Override
+			public void onChanged(Change<? extends Language> c) {
+				while (c.next()) {
+					if (c.wasRemoved()) {
+						key.getValue().getLanguageTokenMap().remove(c.getRemoved());
+					} else if (c.wasAdded()) {
+						for (int i = c.getFrom(); i < c.getTo(); i++) {
+							key.getValue().getLanguageTokenMap().put(c.getList().get(i), "");
+						}
+					}
+				}
+			}
+		});
+		myRootElement.getChildren().add(menuButtonLangs);
+		inputTextProperty().addListener(new ChangeListener<String>() {
+			boolean added;
+			final String badInput = "bad-input-text-field";
+
+			@Override
+			public void changed(ObservableValue<? extends String> observable, String oldValue, String newValue) {
+				newValue = newValue != null ? newValue : "";
+				if (!StringTableKey.idIsProper(newValue)) {
+					if (!added) {
+						added = true;
+						textField.getStyleClass().add(badInput);
+					}
+				} else {
+					added = false;
+					textField.getStyleClass().remove(badInput);
+				}
+				key.setId(newValue);
+			}
+		});
+
+		textField.requestFocus();
+		textField.positionCaret(textField.getText().length());
+
+	}
+
+	@Override
+	protected void ok() {
+		StringTableKey existingKey = table.getKeyById(key.getId());
+		if (existingKey != null) {
+			new KeyAlreadyExistsDialog(key).show();
+			return;
+		}
+		if (key.idIsProper()) {
+			super.ok();
+		} else {
+			textField.requestFocus();
+			beep();
+		}
+	}
+
+	/**
+	 Get the new key. If the dialog was cancelled, will still return a non-null key
+
+	 @return new key
+	 */
+	@NotNull
+	public StringTableKey getKey() {
+		return key;
+	}
+}