+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/stringtable/StringTableEditorPopup.java
@@ -9,6 +9,7 @@ import com.kaylerrenslow.armaDialogCreator.main.ArmaDialogCreator;
 import com.kaylerrenslow.armaDialogCreator.main.Lang;
 import com.kaylerrenslow.armaDialogCreator.util.ValueListener;
 import com.kaylerrenslow.armaDialogCreator.util.ValueObserver;
+import javafx.beans.property.BooleanProperty;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
 import javafx.collections.FXCollections;
@@ -16,6 +17,7 @@ import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.geometry.Insets;
+import javafx.geometry.Orientation;
 import javafx.scene.control.*;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.Priority;
@@ -32,12 +34,67 @@ import java.util.*;
  @since 12/14/2016 */
 public class StringTableEditorPopup extends StagePopup<VBox> {
 
+	private final StringTableEditorTabPane tabPane;
+
+	/** String to be used for when {@link StringTableKey#getPackageName()}==null */
+	private final String noPackageName,
+	/** String to be used for when {@link StringTableKey#getContainerName()}==null */
+	noContainerName;
+
+	private StringTable table;
+
 	public StringTableEditorPopup(@NotNull StringTable table, @NotNull StringTableWriter writer, @NotNull StringTableParser parser) {
 		super(ArmaDialogCreator.getPrimaryStage(), new VBox(0), Lang.ApplicationBundle().getString("Popups.StringTable.popup_title"));
+		this.table = table;
 		ResourceBundle bundle = Lang.ApplicationBundle();
 
-		StringTableEditorTabPane tabPane = new StringTableEditorTabPane(table);
+		noPackageName = bundle.getString("Popups.StringTable.no_package");
+		noContainerName = bundle.getString("Popups.StringTable.no_container");
 
+		Button btnInsert = new Button("", new ImageView(ADCImages.ICON_PLUS));
+		btnInsert.setTooltip(new Tooltip(bundle.getString("Popups.StringTable.Tab.Edit.insert_key_tooltip")));
+		btnInsert.setOnAction(new EventHandler<ActionEvent>() {
+			@Override
+			public void handle(ActionEvent event) {
+				NewStringTableKeyDialog dialog = new NewStringTableKeyDialog(getTable());
+				dialog.show();
+				if (dialog.wasCancelled()) {
+					return;
+				}
+				EditTab editTab = tabPane.getEditTab();
+				StringTableKeyDescriptor newKey = editTab.addNewKey(dialog.getKey());
+				getTable().getKeys().add(dialog.getKey());
+				editTab.getListView().getSelectionModel().select(newKey);
+				editTab.getListView().scrollTo(newKey);
+			}
+		});
+		Button btnRemove = new Button("", new ImageView(ADCImages.ICON_MINUS));
+		btnRemove.setTooltip(new Tooltip(bundle.getString("Popups.StringTable.Tab.Edit.remove_key_tooltip")));
+		btnRemove.setOnAction(new EventHandler<ActionEvent>() {
+			@Override
+			public void handle(ActionEvent event) {
+				ListView<StringTableKeyDescriptor> listView = tabPane.getEditTab().getListView();
+				StringTableKeyDescriptor selected = listView.getSelectionModel().getSelectedItem();
+				if (selected == null) {
+					return;
+				}
+				SimpleResponseDialog dialog = new SimpleResponseDialog(
+						ArmaDialogCreator.getPrimaryStage(),
+						btnRemove.getTooltip().getText(),
+						String.format(bundle.getString("Popups.StringTable.Tab.Edit.remove_key_popup_body_f"), selected.getKey().getId()),
+						true, true, false
+				);
+				dialog.show();
+				if (dialog.wasCancelled()) {
+					return;
+				}
+				tabPane.getEditTab().removeKey(selected);
+			}
+		});
+
+		tabPane = new StringTableEditorTabPane(table, btnRemove.disableProperty(), this);
+
+		btnRemove.setDisable(tabPane.getEditTab().getListView().getSelectionModel().isEmpty());
 		Button btnRefresh = new Button("", new ImageView(ADCImages.ICON_REFRESH));
 		btnRefresh.setTooltip(new Tooltip(bundle.getString("Popups.StringTable.ToolBar.reload_tooltip")));
 		btnRefresh.setOnAction(new EventHandler<ActionEvent>() {
@@ -54,7 +111,8 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 					return;
 				}
 				try {
-					tabPane.setToTable(parser.createStringTableInstance());
+					setTable(parser.createStringTableInstance());
+					tabPane.setToTable(getTable());
 				} catch (IOException e) {
 					e.printStackTrace();
 				}
@@ -67,14 +125,15 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 			@Override
 			public void handle(ActionEvent event) {
 				try {
-					writer.writeTable(table);
+					writer.writeTable(getTable());
 				} catch (IOException e) {
 					e.printStackTrace();
 				}
 			}
 		});
 
-		myRootElement.getChildren().add(new ToolBar(btnRefresh, btnSave));
+		btnRemove.disabledProperty();
+		myRootElement.getChildren().add(new ToolBar(btnRefresh, btnSave, new Separator(Orientation.VERTICAL), btnInsert, btnRemove));
 		myRootElement.getChildren().add(tabPane);
 		VBox.setVgrow(tabPane, Priority.ALWAYS);
 
@@ -82,17 +141,42 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 
 	}
 
+	private void setTable(StringTable table) {
+		this.table = table;
+	}
+
+	private StringTable getTable() {
+		return table;
+	}
+
 	private static class StringTableEditorTabPane extends TabPane {
 		private final ValueObserver<Language> previewLanguageObserver = new ValueObserver<>(KnownLanguage.Original);
+		private final BooleanProperty disableRemove;
+		private final StringTableEditorPopup editorPopup;
+		private EditTab editTab;
 
-		public StringTableEditorTabPane(@NotNull StringTable table) {
+		public StringTableEditorTabPane(@NotNull StringTable table, @NotNull BooleanProperty disableRemove, @NotNull StringTableEditorPopup editorPopup) {
+			this.disableRemove = disableRemove;
+			this.editorPopup = editorPopup;
 			setToTable(table);
 		}
 
 		public void setToTable(@NotNull StringTable table) {
 			getTabs().clear();
-			getTabs().add(new EditTab(table, previewLanguageObserver));
+			editTab = new EditTab(table, previewLanguageObserver, editorPopup);
+			editTab.getListView().getSelectionModel().selectedItemProperty().addListener(new ChangeListener<StringTableKeyDescriptor>() {
+				@Override
+				public void changed(ObservableValue<? extends StringTableKeyDescriptor> observable, StringTableKeyDescriptor oldValue, StringTableKeyDescriptor selected) {
+					disableRemove.setValue(selected == null);
+				}
+			});
+			getTabs().add(editTab);
 			getTabs().add(new ConfigTab(table, previewLanguageObserver));
+		}
+
+		@NotNull
+		public EditTab getEditTab() {
+			return editTab;
 		}
 	}
 
@@ -125,14 +209,23 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 	}
 
 	private static class EditTab extends Tab {
+		private static final Comparator<StringTableKeyDescriptor> comparator = new Comparator<StringTableKeyDescriptor>() {
+			@Override
+			public int compare(StringTableKeyDescriptor o1, StringTableKeyDescriptor o2) {
+				return o1.getKey().getId().compareToIgnoreCase(o2.getKey().getId());
+			}
+		};
+		;
 		private final ObservableList<StringTableKeyDescriptor> listViewItemList;
+		private ValueObserver<Language> previewLanguageObserver;
+		private StringTableEditorPopup editorPopup;
 
 		private final List<StringTableKeyDescriptor> allItems = new LinkedList<>();
 		private final ListView<StringTableKeyDescriptor> lvMatch = new ListView<>();
 		private final StringTableKeyEditorPane editorPane;
 
 
-		public EditTab(@NotNull StringTable table, @NotNull ValueObserver<Language> previewLanguageObserver) {
+		public EditTab(@NotNull StringTable table, @NotNull ValueObserver<Language> previewLanguageObserver, @NotNull StringTableEditorPopup editorPopup) {
 			super(Lang.ApplicationBundle().getString("Popups.StringTable.Tab.Edit.tab_title"));
 
 			listViewItemList = FXCollections.observableList(new ArrayList<>(), new Callback<StringTableKeyDescriptor, javafx.beans.Observable[]>() {
@@ -146,6 +239,8 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 					};
 				}
 			}); //for some reason, can't have a LinkedList as the underlying list implementation if we want the list view to update the displayed cell text automatically
+			this.previewLanguageObserver = previewLanguageObserver;
+			this.editorPopup = editorPopup;
 
 			previewLanguageObserver.addListener(new ValueListener<Language>() {
 				@Override
@@ -156,37 +251,24 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 				}
 			});
 
-			editorPane = new StringTableKeyEditorPane(previewLanguageObserver);
+			editorPane = new StringTableKeyEditorPane(table, previewLanguageObserver);
 
-			setClosable(false);
 			ResourceBundle bundle = Lang.ApplicationBundle();
 
 			lvMatch.setPlaceholder(new Label(bundle.getString("Popups.StringTable.Tab.Edit.Search.no_match")));
 			lvMatch.setStyle("-fx-font-family:monospace");
-			final String noPackageName = bundle.getString("Popups.StringTable.no_package");
-			final String noContainerName = bundle.getString("Popups.StringTable.no_container");
 			for (StringTableKey key : table.getKeys()) {
-				StringTableKeyDescriptor descriptor = new StringTableKeyDescriptor(key, noPackageName, noContainerName);
-				descriptor.setPreviewLanguage(previewLanguageObserver.getValue());
-				allItems.add(descriptor);
-				listViewItemList.add(descriptor);
+				addNewKey(key);
 			}
-			final Comparator<StringTableKeyDescriptor> comparator = new Comparator<StringTableKeyDescriptor>() {
-				@Override
-				public int compare(StringTableKeyDescriptor o1, StringTableKeyDescriptor o2) {
-					return o1.getKey().getId().compareToIgnoreCase(o2.getKey().getId());
-				}
-			};
-			listViewItemList.sort(comparator);
-			allItems.sort(comparator);
+
 			lvMatch.setItems(listViewItemList);
 			lvMatch.getSelectionModel().selectedItemProperty().addListener(new ChangeListener<StringTableKeyDescriptor>() {
 				@Override
 				public void changed(ObservableValue<? extends StringTableKeyDescriptor> observable, StringTableKeyDescriptor oldValue, StringTableKeyDescriptor selected) {
 					if (selected != null) {
-						editorPane.setKey(selected.getKey());
+						editorPane.setKey(selected.getKey(), table);
 					} else {
-						editorPane.setKey(null);
+						editorPane.setKey(null, table);
 					}
 				}
 			});
@@ -198,10 +280,49 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 			vbRoot.setFillWidth(true);
 			vbRoot.setPadding(new Insets(10));
 			setContent(vbRoot);
+			setClosable(false);
 
 		}
 
+		@NotNull
+		public ListView<StringTableKeyDescriptor> getListView() {
+			return lvMatch;
+		}
 
+		/**
+		 Use this instead of adding to {@link ListView#getItems()} with {@link #getListView()}
+
+		 @return the key that was added
+		 */
+		public StringTableKeyDescriptor addNewKey(@NotNull StringTableKey key) {
+			StringTableKeyDescriptor descriptor = new StringTableKeyDescriptor(key, editorPopup.noPackageName, editorPopup.noContainerName);
+			descriptor.setPreviewLanguage(previewLanguageObserver.getValue());
+			allItems.add(descriptor);
+			listViewItemList.add(descriptor);
+			listViewItemList.sort(comparator);
+			allItems.sort(comparator);
+			return descriptor;
+		}
+
+		/** Use this instead of removing from {@link ListView#getItems()} with {@link #getListView()} */
+		public void removeKey(@NotNull StringTableKey key) {
+			StringTableKeyDescriptor match = null;
+			for (StringTableKeyDescriptor descriptor : allItems) {
+				if (descriptor.getKey().equals(key)) {
+					match = descriptor;
+					break;
+				}
+			}
+			if (match == null) {
+				return;
+			}
+			removeKey(match);
+		}
+
+		public void removeKey(@NotNull StringTableKeyDescriptor key) {
+			allItems.remove(key);
+			listViewItemList.remove(key);
+		}
 	}