+++ src/com/kaylerrenslow/armaDialogCreator/gui/notification/Notification.java
@@ -2,14 +2,7 @@ package com.kaylerrenslow.armaDialogCreator.gui.notification;
 
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.SimpleBooleanProperty;
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
-import javafx.geometry.Insets;
-import javafx.geometry.Pos;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.layout.*;
-import javafx.scene.text.Font;
+import javafx.scene.layout.Region;
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -17,14 +10,14 @@ import org.jetbrains.annotations.NotNull;
 
  @author Kayler
  @since 11/16/2016 */
-public class Notification {
+public abstract class Notification {
 
 	public static final long DEFAULT_DURATION = 10 * 1000;
-	private final VBox root = new VBox();
-	private final String notificationTitle;
-	private final String notificationText;
-	private final BooleanProperty showProperty = new SimpleBooleanProperty(true);
-	private long displayDurationMilliseconds;
+	protected final String notificationTitle;
+	protected final String notificationText;
+	protected final BooleanProperty showProperty = new SimpleBooleanProperty(true);
+	protected long displayDurationMilliseconds;
+	protected boolean saveToHistory = true;
 
 	/**
 	 A non-error notification that will last {@link #DEFAULT_DURATION} seconds
@@ -59,43 +52,6 @@ public class Notification {
 		this.notificationTitle = notificationTitle;
 		this.notificationText = notificationText;
 		this.displayDurationMilliseconds = displayDurationMilliseconds;
-
-		root.setStyle("-fx-background-color:-fx-control-inner-background;-fx-border-color:black;-fx-border-width:1px;");
-
-		final Button btnClose = new Button("x");
-		btnClose.setOnAction(new EventHandler<ActionEvent>() {
-			@Override
-			public void handle(ActionEvent event) {
-				setShowing(false);
-			}
-		});
-		final Label lblTitle = new Label(notificationTitle);
-		lblTitle.setFont(Font.font(15));
-		lblTitle.setStyle("-fx-text-fill:white");
-		final BorderPane borderPaneTitle = new BorderPane(null, null, btnClose, null, lblTitle);
-		if (isErrorNotification) {
-			borderPaneTitle.setStyle("-fx-background-color:red;");
-		} else {
-			borderPaneTitle.setStyle("-fx-background-color:-fx-accent;");
-		}
-		borderPaneTitle.setPadding(new Insets(5));
-
-		VBox.setVgrow(borderPaneTitle, Priority.NEVER);
-		root.getChildren().add(borderPaneTitle);
-
-
-		final Label lblText = new Label(notificationText);
-		lblText.setWrapText(true);
-		final StackPane stackPaneContent = new StackPane(lblText);
-		stackPaneContent.setAlignment(Pos.TOP_LEFT);
-		stackPaneContent.setPadding(borderPaneTitle.getPadding());
-		root.getChildren().add(stackPaneContent);
-
-		final double width = 360;
-
-		root.setPrefWidth(width);
-		root.setMaxWidth(width);
-
 	}
 
 	/**
@@ -144,15 +100,26 @@ public class Notification {
 
 	/** Get the JavaFX node that represents notification */
 	@NotNull
-	public Region getContentRoot() {
-		return root;
+	public abstract Region getContentRoot();
+
+	/**
+	 By default, this value is true.
+
+	 @return true if the notification should be stored in {@link Notifications#getPastNotifications()}, false if it shouldn't
+	 */
+	public boolean saveToHistory() {
+		return saveToHistory;
 	}
 
+	/** @see #saveToHistory() */
+	public void setSaveToHistory(boolean saveToHistory) {
+		this.saveToHistory = saveToHistory;
+	}
 
 	@Override
 	public String toString() {
 		return "Notification{" +
-				"root=" + root +
+				"root=" + getContentRoot() +
 				", notificationTitle='" + notificationTitle + '\'' +
 				", notificationText='" + notificationText + '\'' +
 				", displayDurationMilliseconds=" + displayDurationMilliseconds +