+++ CHANGES.md
@@ -1,5 +1,8 @@
 **Added:**
-* none
+* String table editor:
+    * remove selected key
+    * insert new key
+    * redid id renaming (went from inline text field to menu item->popup)
 
 **Changed:**
 * none